Cleanup ListView control
Apply .NET 5.0 ListView API and remove custom logic.
